requirements.txt
Django==1.4
South==0.7.5
anyjson==0.3.1
django-appconf==0.5
django-auth-ldap==1.1
django-debug-toolbar==0.9.4
django-form-utils==0.2.0
django-grappelli==2.3.8
django-picklefield==0.2.1
django-simple-captcha==0.3.4
django-tastypie==0.9.11
